08.06 commit

+++ src/main/java/com/example/board/service/PostService.java
@@ -1,10 +1,14 @@
 package com.example.board.service;
 
 import com.example.board.exception.post.PostNotFoundException;
+import com.example.board.exception.user.UserNotAllowedException;
+import com.example.board.exception.user.UserNotFoundException;
+import com.example.board.model.entity.UserEntity;
 import com.example.board.model.post.Post;
 import com.example.board.model.post.PostPostRequestBody;
 import com.example.board.model.entity.PostEntity;
 import com.example.board.repository.PostEntityRepository;
+import com.example.board.repository.UserEntityRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -17,6 +21,8 @@ import java.util.List;
 public class PostService {
 
     private final PostEntityRepository postEntityRepository;
+
+    private final UserEntityRepository userEntityRepository;
 
 
     public List<Post> getPosts() {
@@ -32,29 +38,49 @@ public class PostService {
 
     }
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
 
-        var postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-        PostEntity save = postEntityRepository.save(postEntity);
-        return Post.from(save);
+        var savedPostEntity = postEntityRepository.save(
+                PostEntity.of(postPostRequestBody.body(), currentUser)
+        );
+
+        return Post.from(savedPostEntity);
 
     }
 
-    public Post updatePost(Long postId, PostPostRequestBody postPostRequestBody) {
+    public Post updatePost(Long postId, PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
         var postEntity = postEntityRepository.findById(postId)
                 .orElseThrow(()-> new PostNotFoundException(postId));
 
-        postEntity.setBody(postPostRequestBody.body());
-        var updatedPostEntity
-                = postEntityRepository.save(postEntity);
-        return Post.from(updatedPostEntity);
+        if (postEntity.getUser().equals(currentUser)) {
+            postEntity.setBody(postPostRequestBody.body());
+            var updatedPostEntity
+                    = postEntityRepository.save(postEntity);
+            return Post.from(updatedPostEntity);
+        }else{
+            throw new UserNotAllowedException();
+        }
+
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentUser) {
         var postEntity = postEntityRepository.findById(postId)
                 .orElseThrow(()-> new PostNotFoundException(postId));
 
-        postEntityRepository.delete(postEntity);
+        if(postEntity.getUser().equals(currentUser)){
+            postEntityRepository.delete(postEntity);
+
+        } else {
+            throw new UserNotAllowedException();
+        }
+
+    }
+
+    public List<Post> getPostByUsername(String username) {
+        UserEntity findEntity = userEntityRepository.findByUsername(username)
+                .orElseThrow(() -> new UserNotFoundException(username));
+        return postEntityRepository.findByUser(findEntity).stream()
+                .map(Post::from).toList();
+
     }
 }
